// frontend/src/services/api/recommendation/enums.ts
export enum RecommendationApiRoutes {
  POPULAR = '/recommendations/popular',
  BEST = '/recommendations/best',
  RECENT = '/recommendations/recent',
  FOR_YOU = '/recommendations/for-you',
  HOME = '/recommendations/home',
  PERSONALIZED = '/recommendations/personalized',
  DISCOVER_WEEKLY = '/recommendations/discover-weekly',
  STATS = '/recommendations/stats',
  SIMILAR = '/recommendations/similar/:trackId',
}